requirements.txt
# Core dependencies
streamlit==1.29.0
pandas==2.0.3
numpy==1.24.3
plotly==5.18.0
altair==5.2.0

# Supabase integration
supabase==2.3.0
python-dotenv==1.0.0

# Data processing
openpyxl==3.1.2
xlrd==2.0.1
xlsxwriter==3.1.9

# Additional visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
httpx==0.25.2
pydantic==2.5.0
python-dateutil==2.8.2

# File handling
aiofiles==23.2.1
watchdog==3.0.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
cryptography==41.0.7